Backend/Chatbot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
botocore==1.38.46
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.3
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
durationpy==0.10
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-ollama==0.3.3
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.1
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.0
openai==1.93.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
posthog==5.4.0
propcache==0.3.2
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0